initial commit
included new 2018a mexfile versions for wola, use them as the reference
for the new complex interleaved api in r2018a. USE INTS INSTEAD OF
MWSIZE! THEY SEEM TO BE PROBLEMATIC NOW.
generated 12, 24, 40 thread version for the wola. appears that 24 is the
sweet spot, 40 seems to be limited by cache speed.
note: this is using 1 processor, yet to implement the dual processor
version!