rewrote testcomplex as dot_xcorr_singleChannel
the old method of threading was found to be causing a lot of slowdowns,
especially for small executions per thread; this is likely because the
sleep function was causing excessive context switching (found by
testing). as such the proper method of starting threads using the
structs is now used. testcomplex is henceforth deprecated but kept for
reference.

TODO: edit dot_xcorr for the multichannel xcorr to also use the new
proper threads.