disabled condition for Dec=N in gpuWola
condition is not necessary (although perhaps the kernel is currently not
optimized for Dec<N)
this is because the current kernel associates outputidx->blockidx and
within that, channelidx->threadidx. the only slowdown should occur when
threads per block is less than fftlen, as the block would have to loop
more than once.